+++ sdk/lib/_internal/js_dev_runtime/private/native_typed_data.dart
@@ -1622,6 +1622,8 @@ final class NativeInt32x4 implements Int32x4 {
     return NativeInt32x4._truncated(t, t, t, t);
   }
 
+  NativeInt32x4.zero() : this._truncated(0, 0, 0, 0);
+
   NativeInt32x4.bool(bool x, bool y, bool z, bool w)
     : this.x = x ? -1 : 0,
       this.y = y ? -1 : 0,