+++ sdk/lib/_internal/js_runtime/lib/typed_data_patch.dart
@@ -174,6 +174,8 @@ class Int32x4 {
   @patch
   factory Int32x4.splat(int value) = NativeInt32x4.splat;
   @patch
+  factory Int32x4.zero() = NativeInt32x4.zero;
+  @patch
   factory Int32x4.bool(bool x, bool y, bool z, bool w) = NativeInt32x4.bool;
   @patch
   factory Int32x4.fromFloat32x4Bits(Float32x4 x) =