+++ sdk/lib/_internal/vm/lib/typed_data_patch.dart
@@ -3981,6 +3981,10 @@ class Int32x4 {
 
   @patch
   @pragma("vm:prefer-inline")
+  factory Int32x4.zero() => Int32x4.splat(0);
+
+  @patch
+  @pragma("vm:prefer-inline")
   factory Int32x4.bool(bool x, bool y, bool z, bool w) {
     return _Int32x4FromBools(x, y, z, w);
   }