+++ sdk/lib/typed_data/typed_data.dart
@@ -2683,8 +2683,8 @@ abstract final class Float32x4 {
 
   /// Creates a `Float32x4` with all values being zero.
   ///
-  /// The created value has the same [Float32x4.x], [Float32x4.y], [Float32x4.z]
-  /// and [Float32x4.w] value, which is the 32-bit floating point zero value.
+  /// The created value has the same [x], [y], [z] and [w] value, which is the
+  /// 32-bit floating point zero value.
   external factory Float32x4.zero();
 
   /// Creates a `Float32x4` with 32-bit float values from the provided bits.
@@ -4344,6 +4344,13 @@ abstract final class Int32x4 {
   /// low 32 bits of [value].
   external factory Int32x4.splat(int value);
 
+  /// Creates an `Int32x4` with all values being zero.
+  ///
+  /// The created value has the same [x], [y], [z] and [w] value, which is the
+  /// 32-bit integer zero value.
+  @Since("3.14")
+  external factory Int32x4.zero();
+
   external factory Int32x4.bool(bool x, bool y, bool z, bool w);
   external factory Int32x4.fromFloat32x4Bits(Float32x4 x);
 
@@ -4782,6 +4789,11 @@ abstract final class Int32x4 {
 abstract final class Float64x2 {
   external factory Float64x2(double x, double y);
   external factory Float64x2.splat(double v);
+
+  /// Creates a `Float64x2` with all values being zero.
+  ///
+  /// The created value has the same [x] and [y] value, which is the 64-bit
+  /// floating point zero value.
   external factory Float64x2.zero();
 
   /// Uses the "x" and "y" lanes from [v].